+++ ichnaea/api/locate/cell.py
@@ -91,7 +91,7 @@ def query_cells(query, lookups, model, raven_client):
             return []
 
         try:
-            load_fields = ('lat', 'lon')  # BBB: 'range'
+            load_fields = ('lat', 'lon', 'radius')
             model_iter = model.iterkeys(
                 query.session,
                 hashkeys,
@@ -110,11 +110,7 @@ def query_areas(query, lookups, model, raven_client):
     if not areaids:  # pragma: no cover
         return []
 
-    if model == CellAreaOCID:
-        load_fields = ('lat', 'lon', 'radius')
-    else:
-        load_fields = ('lat', 'lon')  # BBB: 'range'
-
+    load_fields = ('lat', 'lon', 'radius')
     try:
         areas = (query.session.query(model)
                               .filter(model.areaid.in_(areaids))