+++ ichnaea/api/locate/tests/test_cell.py
@@ -1,5 +1,3 @@
-import unittest2 as unittest
-
 from ichnaea.api.locate.cell import (
     CellPositionSource,
     OCIDPositionSource,
@@ -48,7 +46,6 @@ class TestCellPosition(BaseSourceTest):
         result = self.source.search(query)
         self.check_model_result(result, None)
 
-    @unittest.skip('Skip range/radius check.')
     def test_multiple_cells(self):
         cell = CellFactory()
         cell2 = CellFactory(radio=cell.radio, mcc=cell.mcc, mnc=cell.mnc,
@@ -76,21 +73,19 @@ class TestCellPosition(BaseSourceTest):
                 self.source.should_search(query, ResultList(result)))
             check_db_calls(rw=0, ro=0)
 
-    @unittest.skip('Skip range/radius check.')
     def test_smallest_area(self):
-        area = CellAreaFactory(range=25000)
-        area2 = CellAreaFactory(range=30000, lat=area.lat + 0.2)
+        area = CellAreaFactory(radius=25000)
+        area2 = CellAreaFactory(radius=30000, lat=area.lat + 0.2)
         self.session.flush()
 
         query = self.model_query(cells=[area, area2])
         result = self.source.search(query)
         self.check_model_result(result, area)
 
-    @unittest.skip('Skip range/radius check.')
     def test_minimum_radius(self):
         areas = CellAreaFactory.create_batch(2)
-        areas[0].range = CELLAREA_MIN_ACCURACY - 2000
-        areas[1].range = CELLAREA_MIN_ACCURACY + 3000
+        areas[0].radius = CELLAREA_MIN_ACCURACY - 2000
+        areas[1].radius = CELLAREA_MIN_ACCURACY + 3000
         areas[1].lat = areas[0].lat + 0.2
         self.session.flush()